trimmed files
