+++ frontend/src/util/route_util.js
@@ -30,8 +30,6 @@ const Protected = ({ component: Component, loggedIn, ...rest }) => (
     />
 );
 
-// Use the isAuthenitcated slice of state to determine whether a user is logged in
-
 const mapStateToProps = state => ({ loggedIn: state.session.isAuthenticated });
 
 export const AuthRoute = withRouter(connect(mapStateToProps)(Auth));
